fixed elrond model getting cramped in mobile view via css, new shared AppMarketplace comp thats used in HomeElrond and HomeEVM. homeElrond uses AppMarketplace and removed not used imports due to this reuse and also cleaned home tiles to look moe neat and aligned, homeEVM also used AppMarketplace and removed old code and made it mobile view friendly like HomeElrond and better looking error alert in faucet tile,  #121 latest contract set for goerli
