+++ src/Elrond/custom.css
@@ -73,6 +73,7 @@ body[data-rr-ui-modal-open] .dapp-core-ui-component.modal {
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
+  width: 90%;
 }
 
 .dapp-core-ui-component .media-body {
