+++ src/libs/contactAddresses.js
@@ -24,10 +24,10 @@ const contracts = {
         claimsContractAddress: '',
       },
       goerli: {
-        itheumTokenContractAddress: '0x91ff16CDfeF176b1576E640422C5BA281A242400',
-        ddexContractAddress: '0xBDdb6B94d9B60Ac1D788676a287e8c474D68D44A',
-        dNFTContractAddress: '0xD01A4bCeD9324034db6cb03E50b76F58496F5FB8',
-        claimsContractAddress: '',
+        itheumTokenContractAddress: '0xaC9e9eA0d85641Fa176583215447C81eBB5eD7b3',
+        ddexContractAddress: '0x3be0986AE40de157FAEcfcE3F65b3990b54a5ccb',
+        dNFTContractAddress: '0xd742350dC29171670847Ba68858CC5865d1eC245',
+        claimsContractAddress: '0x14f008d74C42055334A21BFAA7231b4f57E4EFa5',
       },
     },
   },
@@ -72,6 +72,7 @@ export const dNFTContractAddress_Rink = contracts.ethereum[contractsMode].rinkeb
 export const tokenContractAddress_Goerli = contracts.ethereum[contractsMode].goerli.itheumTokenContractAddress;
 export const ddexContractAddress_Goerli = contracts.ethereum[contractsMode].goerli.ddexContractAddress;
 export const dNFTContractAddress_Goerli = contracts.ethereum[contractsMode].goerli.dNFTContractAddress;
+export const claimsContractAddress_Goerli = contracts.ethereum[contractsMode].goerli.claimsContractAddress;
 
 // BSC contracts
 export const tokenContractAddress_testnetBSC = contracts.bsc[contractsMode].itheumTokenContractAddress;
